남녀 성비 오류수정, 로그인&로그아웃 마이페이지 즉각 반영, 별점 오류 수정, 내 후기 보기 추가, 회원가입 로직 수정

+++ app/src/main/java/com/youth/farm_volunteering/data/HomeNonghwalData.kt
@@ -9,7 +9,7 @@ open class HomeNonghwalData() : NonghwalData(), Parcelable {
     constructor(parcel: Parcel) : this() {
         nhIdx = parcel.readValue(Int::class.java.classLoader) as? Int
         idx = parcel.readValue(Int::class.java.classLoader) as? Int
-        star = parcel.readValue(Double::class.java.classLoader) as? Double
+        star = parcel.readValue(Float::class.java.classLoader) as? Float
         price = parcel.readValue(Int::class.java.classLoader) as? Int
         name = parcel.readString()
         period = parcel.readString()
