+++ app/src/main/java/com/youth/farm_volunteering/data/PopularSubData.kt
@@ -10,7 +10,7 @@ open class PopularSubData() : NonghwalData(), Parcelable{
     constructor(parcel: Parcel) : this() {
         nhIdx = parcel.readValue(Int::class.java.classLoader) as? Int
         idx = parcel.readValue(Int::class.java.classLoader) as? Int
-        star = parcel.readValue(Double::class.java.classLoader) as? Double
+        star = parcel.readValue(Float::class.java.classLoader) as? Float
         price = parcel.readValue(Int::class.java.classLoader) as? Int
         name = parcel.readString()
         period = parcel.readString()
